sim-lib: refactors lib to place connection info to its respective node file
